thermal: Typcast THERMAL_CSTATE_INVALID to 32 bit
In order to keep the value of THERMAL_NO_LIMIT
consistent across kernel code and the one read
from device tree, we assign the value 0xffffffff
to THERMAL_CSTATE_INVALID. There is an
inconsistency involved if we keep the value as
-1UL in kernel code since we assign just 1 cell
in the dts files while mentioning values for most
properties. And 1 cell can only have 32 bits.

In this case it was the max and min cooling
device state values. In the bindings doc we have
just one cell allocated for these values and
hence having an invalid value more than 32 bits
does not make much sense since the value read from
the device tree is always 32 bit irrespective of
the architecture.

This change solves the purpose since our aim is
to achieve a larger number which is impractical
for a cooling state point of view.

Bug 1551484

Change-Id: Ifda4af70cb95a3f2a2a77d2266d95addc72e2b6d
Signed-off-by: Arun Swain <arswain@nvidia.com>
Reviewed-on: http://git-master/r/495820