thermal: Fix possible null dereference
In the case where a governor is not enabled in the config files but is still
included/used from the device tree file, the thermal-zone-params holds a valid
governor reference, however the thermal_goernor_list doesn't have that governor.

During registration of zones, when a call is made to find the governor, a null
pointer is returned and subsequently accessed.

This patch fixes it and uses the default governor instead.

Change-Id: I753ef72139b22f0b44f99f90ba1586bd40b43944
Signed-off-by: navneet kumar <navneetk@nvidia.com>
Reviewed-on: http://git-master/r/792618
Reviewed-by: Krishna Sitaraman <ksitaraman@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Muthu Subramaniam <muthus@nvidia.com>