fbcon: Compare variable screen info instead of modes
When trying to match a mode to a variable screen info structure, convert
the mode to a variable screen structure rather than the variable screen
info structure to a mode. The reason why this is necessary is because in
some cases the rounding errors introduced by the conversion of a mode to
variable screen info structures causes the conversion back to a mode not
to yield the same mode as the original.

Consider the following example: A video mode with a 3840x2160 resolution
at 60 Hz. The pixel clock is 533.25 MHz and the total raster area 4000x
2200 pixels. struct fb_var_screeninfo stores the pixel clock as a period
of each clock tick:

	var->pixclock = 1000000000 / 533250 = 1875 ps

The refresh rate of this mode can be computed as follows:

	refresh = 533250000 / (4000 * 2200) = 59 Hz

But struct fb_var_screeninfo does not store the refresh rate, so it must
be reconstructed from the total number of pixels and the pixel clock:

	refresh = (1000000000 / 1875) * 1000 / (4000 * 2200)
	        = 533333000 / 8800000 = 60 Hz

With UHD monitors this will cause the 59 Hz mode to be compared with an
almost identical 60 Hz mode, yet the comparison will still fail and the
modes won't be matched. This in turn can cause the struct display's
->mode field to be NULL, which will later on be dereferenced without
proper checking and cause a crash in fbcon_switch() -> display_to_var()
under the console lock.

Converting the mode to match against to a struct fb_var_screeninfo will
make sure that the same rounding error applies to both variable screen
info structures and allows matching to succeed.

Bug 200130894
Bug 200089256
Bug 200158516

Change-Id: I69ef55ecf0de733f9f4f733173fafd4ac5d99145
Signed-off-by: Thierry Reding <treding@nvidia.com>
Signed-off-by: Naveen Kumar S <nkumars@nvidia.com>
Reviewed-on: http://git-master/r/795373
(cherry picked from commit 86783e70adf479693793de862dcb95cbcd7c25cb)
Reviewed-on: http://git-master/r/929925