pinctrl: tegra: Add config property special-function mode
Sometimes, it is required to change the pins to SFIO mode
from GPIO mode dynamically. It is done via GPIO calls like
gpio_request() and gpio_free() which need to know the
GPIO of the pins.

Add support to achieve this via the pinctrl DT support only
via pincontrol set state.

Based on kernel-4.4 change 44b05051

Change-Id: Ic550035433eaacf1920798978453e7ef702651f3
Signed-off-by: Suresh Mangipudi <smangipudi@nvidia.com>
Reviewed-on: http://git-master/r/1275569
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>