arm64: dts: correct dtc include paths
Correct include paths for device tree file compilation. The current
Makefile is overwriting the DTC_FLAGS and DTCCPP_FLAGS instead of
amending to them. Also, fix the include directory to point to
kernel-4.9 instead of kernel-4.4.

While at it, the dtc compilation is expected to be free-standing so remove
all references to kernel include directories.

Change-Id: I7977b3efb59a1ec015603aa173ef6bd84ef2a16a
Signed-off-by: Timo Alho <talho@nvidia.com>
Reviewed-on: http://git-master/r/1278607
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Venkat Reddy Talla <vreddytalla@nvidia.com>