regmap: irq: Set lockdep subclass for lock
This adds a lockdep subclass for the lock in regmap irq data,
based on the interrupt controller nesting depth of the
controller.

This is required to avoid a lockdep error when two regmap irq
controllers are nested. See the following call chain where
'parent' and 'child' represent two regmap-irq controllers:

irq_set_irq_wake(child_irq, ...)
- regmap_irq_lock(child)
- mutex_lock(child lock)
- regmap_irq_sync_unlock(child)
- irq_set_irq_wake(parent_irq, ...)
- regmap_irq_lock(parent)
- mutex_lock(parent lock)

The two locks here are in the same lock class but different
instances, so the behavior safe but requires a subclass
annotation.

Bug 1859760

Change-Id: I9abc12fe3f6439421cf967a62e3ca132ab85d755
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: http://git-master/r/1282211
Tested-by: Timo Alho <talho@nvidia.com>
Reviewed-by: Timo Alho <talho@nvidia.com>