clk: tegra: DFLL regulator undershoot guard-band
Add a mechanism to enforce the regulator undershoot guard-band when
requesting a new DFLL frequency. It is based on the DFLL limits
mechanism we already have.

The value of guard-band is specified by the device tree property
"nvidia,pmic-undershoot-gb'. It can be overriden by writing to the
pmu_undershoot_gb debugfs file.

Based on work by Aleks Frid <afrid@nvidia.com>

Bug 1832505

Change-Id: Ie7b685adcd65e687abe7eb62e8324c767495e1c7
Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1565471
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Timo Alho <talho@nvidia.com>