soc: tegra: use regulator phandles for DVFS
Rather than relying on the regulator name, use regulator -supply properties
to indicate which regulator is used to supply a given rail. Given that
regulator_get can already do this, if the correct device pointer is supplied,
we can just pass the device pointer rather than the DT node pointer to the
Tegra210 DVFS driver and let regulator_get do the rest. This solution is
compatible with existing DTs where we rely on the regulator name because
the names of the -supply properties are identical to the regulator names
used today. Eg. currently we have a regulator named 'vdd-core'. If a
vdd-core-supply property is supplied, the phandle value will be used to
indicate the regulator rather than the regulator named 'vdd-core'.

Bug 200259095

Change-Id: Idb891b7f84663463f0896b3ba1f6a451922cc542
Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-on: http://git-master/r/1306264
Reviewed-by: Timo Alho <talho@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1566225
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Tested-by: Sachin Nikam <snikam@nvidia.com>