ata: ahci_tegra: fix powergate refcount
In powergating framework powegate refcount is set to 1 if
SATA partition is already powered. In AHCI driver it was
unconditionally calling unpowergate which was again
incrementing the refcount. This causes mismatch because
when AHCI calls powergate it was not actually getting
powergated because refcount was still set to 1.

Bug 200294079

Change-Id: Ie17a2b590b86b3bd81ada053adee16722334919e
Signed-off-by: Preetham Chandru R <pchandru@nvidia.com>
Reviewed-on: http://git-master/r/1459012
(cherry picked from commit 6cc9d7765bb5f917b7dfa21425649a292291950d)
Reviewed-on: http://git-master/r/1499121
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Venu Byravarasu <vbyravarasu@nvidia.com>