pwm: tegra: Add support to lower pwm frequency
The PWM output signal depends on the clock source frequency
and PWM divider for period. There is possibility to have lower
PWM frequency which is not possible with higher PWM clock source
and on this case, it is required to have more fine divider of clock
source to get desired lower side frequency.

Add DT property "pwm-minimum-frequency-hz" to findout the platform
specific requirement and adjust the clock source frequency to
support this lower side frequency.

bug 1927065

Change-Id: Icb642d6886b39348b9b22310b9f0446be8087ae5
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1493946
Tested-by: R Raj Kumar <rrajk@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>