soc/tegra: fuse: disable fuse clkon always
Disable keeping fuse clock always enabled and this is done
for chip later to Tegra186 and as part of this add new compatible
for T194. This helps to save power. On earlier chips, fuse clock is
always enabled as there was issue with MCE not enabling the fuse clock
while accessing fuse registers

Bug 200295264

Change-Id: I8e4b85d7e66f9f5dfc3708820227d12ac2053ae8
Signed-off-by: Shardar Shariff Md <smohammed@nvidia.com>
Reviewed-on: http://git-master/r/1493046
GVS: Gerrit_Virtual_Submit
Reviewed-by: svccoveritychecker <svccoveritychecker@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>