dmaengine: tegra-apb: Fix compilation issue
The Tegra APB DMA driver fails to compile because the variable 'i' is
not defined in tegra_dma_runtime_resume(). This was caused by a recent
change to fix the APB system suspend and was not caught because the
driver is not being built for k4.9.

Bug 1936074

Change-Id: Iffa833c0c528e7cbc4ee6de5521cd79fe7b4b1cd
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1496108
GVS: Gerrit_Virtual_Submit
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>