platform: tegra: avoid tegra_bpmp_send()
tegra_bpmp_send() is used only by this driver. Additionally, this will
work only on t21x that uses shared semaphore token based mail
operations.

tegra_bpmp_send() was designed as a "one way" message to bpmp firmware
if the calling driver want to avoid waiting for the response (because it
does not care about the response). At the time of design, it was thought
that this sort of functionality will be needed more often than it is
used today. Implementation of this API requires somewhat unusual
semantics at the handler side (=bpmp firmware) where it touches bits
that are otherwise modified by the kernel.

Due to these factors, replace tegra_bpmp_send() with
tegra_bpmp_send_receive_atomic(). The side effect is the extra wait time
(waiting for bpmp to completely process the message). This is expected
to be negligible.

Change-Id: Ida37ccc07d143cd81bed339c382baf805a36a13f
Signed-off-by: Sivaram Nair <sivaramn@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1516525
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Matthew Longnecker <mlongnecker@nvidia.com>
GVS: Gerrit_Virtual_Submit