clk: tegra: Support DFLL tuning override from DT
To do this we duplicate a number of cvb structure fields in the
tegra_dfll_soc_data structure and initialize them either from the
provided DT values or from the CVB table values if the corresponding
DT property is missing. Also adjust the code to use the tegra_dfll_soc_data
fields instead of the cvb fields.

Change-Id: Ie3427d6fea147ef57a7b406584a9ea80c5321c3e
Signed-off-by: Alex Frid <afrid@nvidia.com>
Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1577132
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>