clk: tegra: dfll: support PWM regulator control
The DFLL can directly generate a PWM signal to control the regulator rather
then sending i2c messages. This patch adds support for this mode. In this mode
the LUT is not used and also there is no regulator object involved because
there is no way to force the regulator voltage without also changing the DFLL
output frequency. Also the register debugfs file is slightly reworked to only
show the i2c registers when i2c mode is in use.

Change-Id: Iaba67b65e44e71428fc934798d975e7b97c74821
Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1553351
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Timo Alho <talho@nvidia.com>