Revert "memory/tegra: don't build upstream MC controller"
As the T210 emc scaling driver is added to the directory,
it should be added back to build this scaling driver.

Meanwhile, there is a downstream TEGRA_MC driver already
under the nvidia repo. So this patch tries to add this
directory back but mutes this upstream mc driver in order
to get rid of warning:
  Error: Driver 'tegra-mc' is already registered, aborting...

Bug 1874686

This reverts commit a2666cd33decf38ee1ec2c7891736d2213993cd6.

Change-Id: I913ee3cd173c2b4f5950a6072f70d2673922f609
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1564306
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>