tegra: powergate: avoid use of CBUS clocks
vic and nvenc power gating sequences are using CBÚS variant of the
module clock during powergate operation. Align the operation with rest
of the power domains that would have cbus clock but are not using
it. The plain version works in this case as cbus is needed only for
rate control, not for enable/disable.

Bug 1874686

Change-Id: I7d14c91efc3b38dd1861dfd17d50612a989d05c3
Signed-off-by: Timo Alho <talho@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1565293
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>