clk: tegra: Do not register pll_d_dsi_out twice
The pll_d_dsi_out clock is registered in tegra210_periph_clk_init()
and tegra210b01_pll_init(), and both of them are called on T210b01
platforms. So a clock -EEXIST error is observed on such platforms:
    [    0.000000] Tegra clk 307: register failed with -17

So this patch moves one of registerations to tegra210_pll_init() as
Kernel 4.4 does to fix this problem.

Bug 200347360

Change-Id: Ibc9a650e254a9c78fd4c8e456470066b7daa1dbb
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1599019
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>