cpufreq: kick start cpufreq stats collection
Ideally, the requested cpufreq should be one of the entries in
cpufreq table. Unfortunately, that can be false when system
initially boots up, in which case, the last_index will be -1
since the initial frequency is not in freq table.

The consequence is that cpufreq_stats will never be updated.

To kick start stats collection, start from the first frequency
that can be found in frequency table if last_index == -1. Update
stats entry, then stats collection is started.

Change-Id: Ia1f7439fdfc336faf9dfda75a9c2e0a69d08918f
Signed-off-by: Bo Yan <byan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1610624
GVS: Gerrit_Virtual_Submit
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>