soc: tegra: Fix T21x GPU DVFS
On K4.9 of_property_for_each_phandle_with_args macro is broken, and
cannot be used (macro introduces for-loop that is not initialized and
not terminated properly with changes in kernel of-interfaces). As a
result T21x GPU DT nodes parsing failed, GPU thermal DVFS is broken
as well. Fixed, by using phandle-with-args parsing of-interfaces
directly, instead of macro.

Change-Id: I52dd9763f505c97bb5923b57e05b50b9225dfd47
Signed-off-by: Alex Frid <afrid@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1669898
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bo Yan <byan@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>