dma-iommu: iommu: arm-smmu: Add ignore_align
Add the ability for the smmu driver to request that memory regions
do not need to be aligned to be passed to the driver.

Normally the dma-iommu layer will only pass aligned regions
to the driver. This causes slowdowns in mapping because
the arm-smmu driver is built to map all at once regardless of
alignment.

Bug 2028058

Change-Id: Ie65f766fe5a4c77c10767bc623fcd17396e015e8
Signed-off-by: David Gilhooley <dgilhooley@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1674453
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>