usb: host: xhci: handle invalid cstream case
If Stop Endpoint command is sent for a stream endpoint which CStream is
not valid, a Transfer Event generated by it will report '0' in the TRB
Pointer and results in below error message from xhci driver:

ERROR Transfer event for disabled endpoint or incorrect stream ring

This is valid case specified in xhci spec. So error message is removed.

Bug 2037969

Change-Id: Ic6a70e3f22e212e8c24ded39878ccaea2146dd6c
Signed-off-by: Henry Lin <henryl@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1681814
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Jim Lin <jilin@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: ChihMin Cheng <ccheng@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>