clk: tegra210b01: Do not init clock rate of pll_d2
The pll_d2 clock is used for tegradc -- display on T210. This
clock is usually configured in the bootloader to support early
display.

When kernel takes over the framebuffer from the bootloader, it
should keep the same clock rate for pll_d2 in order to remain
the same pixel clock.

So this patch removes the pll_d2 from the t210_init_table. The
display driver should take the responsibility to configure the
rate according to the display resolution for initialization.

Bug 200369159
Bug 200375996

Change-Id: Ie3b8a9bc9cb671fb1dd81da3125835c799940f45
Signed-off-by: David DSH <ddastoussthi@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1691503
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
Reviewed-by: Prafull Suryawanshi <prafulls@nvidia.com>
Reviewed-by: Santosh Galma <galmar@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>