PCI: designware: rename Kconfig symbol
A future change will backport a new DWC PCIe driver from kernel 4.14.
Unfortunately, this causes a Kconfig symbol collision, which in turn
causes parts of both drivers to always be built when either is enabled,
which causes link failures. Solve this by renaming the existing Kconfig
symbol to a legacy name. Eventually, we'll get rid of the current driver
and replace it with the new one, so renaming the current driver causes
the least amount of churn.

Bug 2312059

Change-Id: I4de0e442813f58868f20883d1290a87f6dbe9385
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1962219
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Manikanta Maddireddy <mmaddireddy@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>