platform: tegra: pm_domains: check initial power status at initial time
Currently, the genpd depends on the "is_off" property in the DT to know
the initial power status from bootloader to kernel. But it could cause
conflict if different bootloaders remain different power status for the
kernel. So changing it as runtime detection not rely on the "is_off"
property.

Bug 200492630

Change-Id: I515c1697202b1e6e3e5fbf37d1cebc6c6abb223d
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2017981
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>