PCI: tegra: Fix speed change code
Speed change to Gen-2 should take place when there is no traffic on the
bus and the way software can make sure that is to attempt speed change to
Gen-2 before registering the devices with PCIe sub-system. This patch
does exactly that. It also fixes link retrain timeout value to 100000us.

Bug 200519254

Change-Id: I820a78d907a25e25e1e95e22456793d5e33ad833
Signed-off-by: Vidya Sagar <vidyas@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2119203
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>