max17042_battery: provide accurate battery status
Before, the driver would only update capacity when capacity property was called.
That value is used also in status property.
Additionally, it was also unconditionally checked.
There would be cases where the driver would report that the battery was charging while full or full while discharging.

This commit alleviates the issues by updating capacity when charger reports full and only allows status full to be reported if battery is charging or full.