HID: nintendo: Optimize rumble amplitudes for special effects
Now strong and weak force is used for left and right joycon respectively.
Both low and high frequencies are used to create a mixed effect that slightly resembles strong/weak effects.
Additionally, the max amplitudes are tuned for user experience, in order to avoid huge audible and annoying sounds and numbness to hands.

Finally, the worqueue is now ordered to avoid executing multiple rumble commands simultaneously.

Signed-off-by: CTCaer <ctcaer@gmail.com>